# target/classes/openapi.yaml
openapi: 3.1.1
info:
  title: NewsNow Image API
  description: API to create image resizing tasks and query their status
  version: 1.0.0

servers:
  - url: http://{host}:{port}/{basePath}
    description: Local server
    variables:
      host:
        default: localhost
        description: A user-specific subdomain. Use 'localhost' to work on a local computer.
      port:
        enum:
          - '8080'
          - '443'
        default: '8080'
      basePath:
        default: back

tags:
  - name: tasks
    description: Operations in image tasks

paths:
  /task:
    post:
      tags: [tasks]
      summary: Create an image resize task
      description: >
        Receives an image with specified dimensions. Returns the metadata for the created task.
        Optionally, the implementation can store the resized image and return its URL.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, width, height]
              properties:
                file:
                  type: string
                  format: binary
                  description: Original image to resize
                width:
                  type: integer
                  format: int32
                  minimum: 1
                  description: Desired width in pixels
                height:
                  type: integer
                  format: int32
                  minimum: 1
                  description: Desired height in pixels
            encoding:
              file:
                contentType: image/png, image/jpeg, image/webp
      responses:
        "201":
          description: Task created
          headers:
            Location:
              description: URL of the created resource (/task/{taskId})
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request (missing file or incorrect dimensions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error while processing the image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /task/{taskId}:
    get:
      tags: [tasks]
      summary: Get task information by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /task/all:
    get:
      tags: [tasks]
      summary: Get all task information
      responses:
        "200":
          description: Information about all tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "500":
          description: Internal server problems
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Task:
      type: object
      required: [id, createdAt, md5, resolution, status]
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        createdAt:
          type: string
          format: date-time
          description: Task creation time (UTC)
        md5:
          type: string
          description: MD5 fingerprint of the original file (hex)
          pattern: "^[A-Fa-f0-9]{32}$"
          example: "d41d8cd98f00b204e9800998ecf8427e"
        resolution:
          type: string
          description: Resolution applied to the result (width x height)
          example: "800x600"
        imageUrl:
          type: string
          nullable: true
          description: URL of the resized image (if stored)
          example: "https://localhost/images/imgExample-800x600.jpg"

    Error:
      type: object
      required: [message]
      properties:
        localDateTime:
          type: string
          format: date-time
        message:
          type: string
        code:
          type: int
          nullable: true